+++ slang/instructor/src/label_converter.rs
@@ -1,3 +1,3 @@
 pub trait LabelConverter {
-    fn offset_of(&self, label_name: &str) -> Option<u32>;
+    fn offset_of(&self, label_name: &str) -> Option<u16>;
 }
